+++ src/AppBundle/DataFixtures/ORM/LoadExampleUsers.php
@@ -2,40 +2,49 @@
 
 namespace Raddit\AppBundle\DataFixtures\ORM;
 
-use Doctrine\Common\DataFixtures\FixtureInterface;
-use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
+use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\Persistence\ObjectManager;
 use Raddit\AppBundle\Entity\User;
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 
-class LoadExampleUsers implements FixtureInterface, OrderedFixtureInterface, ContainerAwareInterface {
+class LoadExampleUsers extends AbstractFixture implements ContainerAwareInterface {
     use ContainerAwareTrait;
 
     /**
      * {@inheritdoc}
      */
     public function load(ObjectManager $manager) {
-        $user1 = new User();
-        $user1->setUsername('emma');
-        $user1->setPassword($this->container->get('security.password_encoder')->encodePassword($user1, 'goodshit'));
-        $user1->setEmail('emma@example.com');
-        $user1->setAdmin(true);
-        $manager->persist($user1);
+        $encoder = $this->container->get('security.password_encoder');
 
-        $user2 = new User();
-        $user2->setUsername('zach');
-        $user2->setPassword($this->container->get('security.password_encoder')->encodePassword($user2, 'example2'));
-        $user2->setEmail('zach@example.com');
-        $manager->persist($user2);
+        foreach ($this->provideUsers() as $data) {
+            $user = new User();
+            $user->setUsername($data['username']);
+            $user->setPassword($encoder->encodePassword($user, $data['password']));
+            $user->setAdmin($data['admin']);
+            $user->setEmail($data['email']);
+
+            $this->addReference('user-'.$data['username'], $user);
+
+            $manager->persist($user);
+        }
 
         $manager->flush();
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getOrder() {
-        return 0;
+    private function provideUsers() {
+        yield [
+            'username' => 'emma',
+            'password' => 'goodshit',
+            'email' => 'emma@example.com',
+            'admin' => true,
+        ];
+
+        yield [
+            'username' => 'zach',
+            'password' => 'example2',
+            'email' => 'zach@example.com',
+            'admin' => false,
+        ];
     }
 }